Fix for crash on http connection close
Encountered an issue where I would experience an uncaught exception (see below stack) approximately two minutes after issuing the first request to edgecast. The underlying cause appears to be the JSONClient leaving the connection open after the request is complete and then treating the connection close as an error. My proposed solution would be to set the "agent" option on the JSONClient object to false, forcing the connection to explicitly close after the request is complete.

Error: read ECONNRESET
    at errnoException (net.js:901:11)
    at TCP.onread (net.js:556:19)